Separating async views from the Dashboard app
