+++ dashboard/urls.py
@@ -1,13 +1,13 @@
 from django.urls import path
 
-from . import views
+from . import views, ajax
 
 app_name = 'dashboard'
 
 urlpatterns = [
     path('', views.index, name='index'),
     path('charts/', views.charts, name='charts'),
-    path('charts/ajax/', views.charts_ajax, name='charts_ajax'),
+    path('charts/ajax/', ajax.charts_async, name='charts_ajax'),
     path('products/', views.products, name='products'),
     path('admin-prod/', views.admin_products, name='admin_prod'),
     path('vip-prod/', views.vip_products, name='vip_prod'),
@@ -15,8 +15,8 @@ urlpatterns = [
     path('profile/', views.user_profile, name="user_profile"),
     path('activity_log/', views.activity_log, name="activity_log"),
     path('search/', views.search, name="search"),
-    path('search/ajax/', views.search_ajax, name="search_ajax"),
-    path('email/send/', views.send_email_async, name="email_ajax"),
+    path('search/ajax/', ajax.search_async, name="search_ajax"),
+    path('email/send/', ajax.send_email_async, name="email_ajax"),
     path('signup/', views.signup, name='signup'),
     path('login/', views.CustomLoginView.as_view(), name='login'),
     path('logout/', views.CustomLogoutView.as_view(), name='logout'),
